# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Travel Audio Guide API
  version: "1.0.0"
  description: |
    ## Instructions
    This is the backend API for the travel audio guide.

    ## Verification Method
    All `/api/v1/*` requests require a Supabase JWT token.

    **Steps:**
    1. Click the **Authorize** button in the upper right corner.
    2. Enter the Supabase `access_token` in the `bearerAuth (http, Bearer)` field.
    3. Click **Authorize** → **Close**.
    4. Subsequent requests will automatically include `Authorization: Bearer <token>`.

    ## Security Design
    - The frontend **does not pass user_id**; the backend automatically obtains it from the JWT `claims.sub`.
    - All user-scoped SQL queries use the user_id parsed from the JWT as a condition.
    - `/open-api/*` routes are public tourism data and intentionally do not require authentication.
servers:
  - url: http://localhost:8080
    description: Local development environment
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The access token obtained after logging into Supabase.
        When entering it, **you do not need to** add the `Bearer` prefix; Swagger UI will add it automatically.
  schemas:
...
    Profile:
      type: object
      description: User profile information
      properties:
        id:
          type: string
          format: uuid
          description: "The Supabase user_id comes from JWT claims.sub and is filled in by the backend."
          example: "0e9b1a0e-1234-5678-abcd-xxxxxxxxxxxx"
        email:
          type: string
          format: email
          nullable: true
          description: "login email"
          example: "user@example.com"
        display_name:
          type: string
          nullable: true
          description: "Display name, 1–50 characters"
          example: "Sun"
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: "Profile picture URL"
          example: "https://example.com/avatar.png"
        preferred_language:
          type: string
          description: "Preferred language"
          enum: [zh-TW, zh-CN, en, ja]
          example: "zh-TW"
        created_at:
          type: string
          format: date-time
          example: "2026-06-14T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-14T00:10:00Z"

    # UpdateMeRequest (Body for PUT /me)
    UpdateMeRequest:
      type: object
      description: |
        Request body for updating the profile.
        **All fields are optional**; only send the fields you want to update, and omitted fields keep their original values.
        **Do not send user_id**; the backend retrieves it automatically from the JWT.
      properties:
        display_name:
          type: string
          nullable: true
          minLength: 1
          maxLength: 50
          description: "Display name, 1-50 characters"
          example: "Sun"
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: "Avatar image URL, must be a valid URL (https://...)"
          example: "https://example.com/avatar.png"
        preferred_language:
          type: string
          nullable: true
          enum: [zh-TW, zh-CN, en, ja]
          description: "Preferred language, allowed values: zh-TW / zh-CN / en / ja"
          example: "zh-TW"
      # Note: required is intentionally omitted; all fields are optional

    # Standard response format
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Profile'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: "Error message"
          example: "missing authorization header"

    # Taipei Travel Attractions compatibility response
    AttractionsResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 471
        data:
          type: array
          items:
            $ref: '#/components/schemas/Attraction'
...
    AttractionsErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "upstream service unavailable"

    # Taipei Travel Events compatibility responses
    NewsResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 35
        data:
          type: array
          items:
            $ref: '#/components/schemas/News'
    News:
      type: object
      required:
        - id
        - title
        - description
        - begin
        - end
        - posted
        - modified
        - url
        - files
        - links
      properties:
        id:
          type: integer
          example: 45417
        title:
          type: string
          example: "水門何時關？北市府設多管道　民眾防汛資訊不漏接"
        description:
          type: string
        begin:
          type: string
          nullable: true
          example: null
        end:
          type: string
          nullable: true
          example: null
        posted:
          type: string
          example: "2023-04-04 09:06:00 +08:00"
        modified:
          type: string
          example: "2023-04-10 10:21:08 +08:00"
        url:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/EventFile'
        links:
          type: array
          items:
            $ref: '#/components/schemas/EventLink'
    ActivityResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 24
        data:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
...
    CalendarResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 23
        data:
          type: array
          items:
            $ref: '#/components/schemas/Calendar'
...
    EventFile:
      type: object
      properties:
        src:
          type: string
        subject:
          type: string
        ext:
          type: string
    EventLink:
      type: object
      properties:
        src:
          type: string
        subject:
          type: string
    EventsErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "upstream service unavailable"

    # Taipei Travel Media compatibility response
    AudioResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 140
        data:
          type: array
          items:
            $ref: '#/components/schemas/Audio'
    Audio:
      type: object
      required:
        - id
        - title
        - summary
        - url
        - file_ext
        - modified
      properties:
        id:
          type: integer
          example: 28
        title:
          type: string
          example: "北投圖書館"
        summary:
          type: string
          nullable: true
          example: null
        url:
          type: string
          example: "https://www.travel.taipei/audio/28"
        file_ext:
          type: string
          nullable: true
          example: null
        modified:
          type: string
          example: "2025-12-10 15:55:41 +08:00"
    MediaErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "unsupported language: xx"

    # Taipei Travel Tours compatibility response
    ThemeResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 87
        data:
          type: array
          items:
            $ref: '#/components/schemas/Theme'
...
    ToursErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "unsupported language: xx"

    # Taipei Travel Miscellaneous compatibility response
    CategoriesResponse:
      type: object
      required:
        - total
        - data
      properties:
        total:
          type: integer
          example: 20
        data:
          $ref: '#/components/schemas/CategoriesData'
    CategoriesData:
      type: object
      description: >
        Confirmed for type=Tours only. The wrapper key is "Category"
        (singular, capitalized). Other "type" values have not yet been
        verified to use the same key.
      required:
        - Category
      properties:
        Category:
          type: array
          items:
            $ref: '#/components/schemas/Category'
...
    MiscellaneousErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "invalid parameter: type (Foo)"
paths:
...
  /api/v1/me:
    get:
      tags: [Me]
      summary: Get the profile of the currently logged-in user
      description: |
        Retrieve the user_id from the JWT token and query the corresponding profile.

        ### Special Notes
        - **The profile will be automatically created if it does not exist.** The frontend does not need to call the profile creation API separately.
        - For first login, simply call this API.

        ### Required Headers
        ```
        Authorization: Bearer <supabase_access_token>
        ```

        ### Backend Flow
        1. Retrieve the Bearer token from the `Authorization` header.
        2. Verify the Supabase JWT (HS256).
        3. Retrieve the `user_id` from `claims.sub`.
        4. Query the `profiles` table.
        5. If it does not exist, it will be automatically created. `preferred_language` is preset to `zh-TW`.
        6. Return the profile.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully obtained (or automatically created) profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                existing_profile:
                  summary: A profile already exists.
                  value:
                    success: true
                    data:
                      id: "0e9b1a0e-1234-5678-abcd-xxxxxxxxxxxx"
                      email: "user@example.com"
                      display_name: "Sun"
                      avatar_url: "https://example.com/avatar.png"
                      preferred_language: "zh-TW"
                      created_at: "2026-06-14T00:00:00Z"
                      updated_at: "2026-06-14T00:10:00Z"
                new_profile:
                  summary: The profile is automatically created upon first login.
                  value:
                    success: true
                    data:
                      id: "0e9b1a0e-1234-5678-abcd-xxxxxxxxxxxx"
                      email: "user@example.com"
                      display_name: null
                      avatar_url: null
                      preferred_language: "zh-TW"
                      created_at: "2026-06-14T00:00:00Z"
                      updated_at: "2026-06-14T00:00:00Z"
        "401":
          description: |
            Unauthorized, possible reasons:
            - No `Authorization` header
            - Incorrect format (not `Bearer <token>`)
            - Token expired
            - Incorrect token signature (incorrect JWT Secret)
            - Token's `role` is not `authenticated`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_header:
                  summary: No Authorization header included
                  value:
                    success: false
                    error: "missing authorization header"
                invalid_token:
                  summary: The token is invalid or has expired.
                  value:
                    success: false
                    error: "invalid or expired token"
        "500":
          description: Server internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "failed to get profile"

    put:
      tags: [Me]
      summary: Update the current logged-in user's profile
      description: |
        Update the profile data of the currently logged-in user.

        ### Important Rules
        - **All fields are optional**; only send the fields you want to update, and omitted fields keep their original values
        - **Do not send `user_id`**; the backend automatically gets it from JWT `claims.sub`
        - Call `GET /api/v1/me` first to ensure the profile exists before updating

        ### Update Name Only (Leave Other Fields Unchanged)
        ```json
        { "display_name": "Sun" }
        ```

        ### Update Language Only (Leave Other Fields Unchanged)
        ```json
        { "preferred_language": "en" }
        ```

        ### Update All Fields
        ```json
        {
          "display_name": "Sun",
          "avatar_url": "https://example.com/avatar.png",
          "preferred_language": "zh-TW"
        }
        ```

        ### Field Rules
        | Field | Rule |
        |------|------|
        | `display_name` | 1-50 characters; leading/trailing spaces are automatically trimmed |
        | `avatar_url` | Must be a valid URL (https://...) |
        | `preferred_language` | Allowed values only: `zh-TW` `zh-CN` `en` `ja` |
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
            examples:
              name_only:
                summary: Update name only
                value:
                  display_name: "Sun"
              language_only:
                summary: Update language only
                value:
                  preferred_language: "en"
              update_all:
                summary: Update all fields
                value:
                  display_name: "Sun"
                  avatar_url: "https://example.com/avatar.png"
                  preferred_language: "zh-TW"
      responses:
        "200":
          description: Update succeeded; returns the full updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  id: "0e9b1a0e-1234-5678-abcd-xxxxxxxxxxxx"
                  email: "user@example.com"
                  display_name: "Sun"
                  avatar_url: "https://example.com/avatar.png"
                  preferred_language: "zh-TW"
                  created_at: "2026-06-14T00:00:00Z"
                  updated_at: "2026-06-14T00:10:00Z"
        "400":
          description: Invalid request body format (not valid JSON)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "invalid JSON body"
        "401":
          description: Unauthorized (token is invalid or expired)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "missing authorization header"
        "403":
          description: Token role is not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "token role is not authenticated"
        "404":
          description: |
            Profile does not exist.
            Please call `GET /api/v1/me` first to create the profile, then update it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "profile not found, call GET /api/v1/me first"
        "422":
          description: |
            Field validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_name:
                  summary: Invalid name length
                  value:
                    success: false
                    error: "display_name: must be 1–50 characters"
                invalid_url:
                  summary: Invalid URL format
                  value:
                    success: false
                    error: "avatar_url: must be a valid URL (https://...)"
                invalid_language:
                  summary: Language is not in the allowed list
                  value:
                    success: false
                    error: "preferred_language: must be one of: zh-TW, zh-CN, en, ja"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "failed to update profile"

  # Taipei Travel attractions proxy
  /open-api/{lang}/Attractions/All:
    get:
      tags:
        - Attractions
      summary: List attractions
      description: |
        Compatibility proxy for the Taipei Travel open API.

        No authentication is required because this endpoint serves
        public tourism data.

        The endpoint, query parameters, JSON field names, field types,
        and response structure are kept compatible with the upstream API.

        The Go backend internally uses its own DTO and mapper layer,
        while preserving the contract expected by existing clients.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
              - ja
              - ko
          example: zh-tw
        - name: categoryIds
          in: query
          required: false
          description: >
            Comma-separated attraction category IDs.
            Example: 13,15
          schema:
            type: string
          example: "13"
        - name: nlat
          in: query
          required: false
          description: Latitude in WGS84 for nearby attraction queries
          schema:
            type: number
            format: double
          example: 25.04356
        - name: elong
          in: query
          required: false
          description: Longitude in WGS84 for nearby attraction queries
          schema:
            type: number
            format: double
          example: 121.51436
        - name: page
          in: query
          required: false
          description: Page number; upstream returns up to 30 results per page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: Attractions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionsResponse'
        "400":
          description: Unsupported language
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionsErrorResponse'
              example:
                error: "unsupported language: xx"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionsErrorResponse'
              example:
                error: "upstream service unavailable"

  # Taipei Travel events proxy
  /open-api/{lang}/Events/News:
    get:
      tags:
        - Events
      summary: List tourism news
      description: |
        Compatibility proxy for the Taipei Travel Events/News open API.

        No authentication is required because this endpoint serves
        public tourism news.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
              - ja
              - ko
          example: zh-tw
        - name: begin
          in: query
          required: false
          description: Start date, format yyyy-MM-dd
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          description: End date, format yyyy-MM-dd
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          description: Page number; upstream returns up to 30 results per page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: News retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsResponse'
        "400":
          description: Unsupported language
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsErrorResponse'
              example:
                error: "unsupported language: xx"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsErrorResponse'
              example:
                error: "upstream service unavailable"
  /open-api/{lang}/Events/Activity:
    get:
      tags:
        - Events
      summary: List activities and exhibitions
      description: |
        Compatibility proxy for the Taipei Travel Events/Activity open API.

        No authentication is required because this endpoint serves
        public tourism activity data.

        Note that `nlat` and `elong` in the response are returned as strings
        by the upstream API, unlike the Attractions endpoint where they are numbers.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
              - ja
              - ko
              - es
              - id
              - th
              - vi
          example: zh-tw
        - name: begin
          in: query
          required: false
          description: Start date, format yyyy-MM-dd
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          description: End date, format yyyy-MM-dd
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          description: Page number; upstream returns up to 30 results per page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: Activities retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        "400":
          description: Unsupported language
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsErrorResponse'
              example:
                error: "unsupported language: xx"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsErrorResponse'
              example:
                error: "upstream service unavailable"
  /open-api/{lang}/Events/Calendar:
    get:
      tags:
        - Events
      summary: List event calendar entries
      description: |
        Compatibility proxy for the Taipei Travel Events/Calendar open API.

        No authentication is required because this endpoint serves
        public tourism event calendar data.

        Only three languages are supported by the upstream API for this
        endpoint: zh-tw, zh-cn, en.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
          example: zh-tw
        - name: categoryId
          in: query
          required: false
          description: Category ID
          schema:
            type: integer
            format: int64
        - name: begin
          in: query
          required: false
          description: Start date, format yyyy-MM-dd
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          description: End date, format yyyy-MM-dd
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          description: Page number; upstream returns up to 30 results per page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: Calendar entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarResponse'
        "400":
          description: Unsupported language
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsErrorResponse'
              example:
                error: "unsupported language: xx"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsErrorResponse'
              example:
                error: "upstream service unavailable"

  # Taipei Travel media proxy
  /open-api/{lang}/Media/Audio:
    get:
      tags:
        - Media
      summary: List audio guide entries
      description: |
        Compatibility proxy for the Taipei Travel Media/Audio open API.

        No authentication is required because this endpoint serves
        public audio guide metadata.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
              - ja
              - ko
          example: zh-tw
        - name: page
          in: query
          required: false
          description: Page number; upstream returns up to 30 results per page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: Audio guide entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioResponse'
        "400":
          description: Unsupported language
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaErrorResponse'
              example:
                error: "unsupported language: xx"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaErrorResponse'
              example:
                error: "upstream service unavailable"

  # Taipei Travel tours proxy
  /open-api/{lang}/Tours/Theme:
    get:
      tags:
        - Tours
      summary: List themed tours
      description: |
        Compatibility proxy for the Taipei Travel Tours/Theme open API.

        No authentication is required because this endpoint serves
        public tourism data.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
              - ja
              - ko
          example: zh-tw
        - name: categoryIds
          in: query
          required: false
          description: >
            Comma-separated category IDs, e.g. "532,228". Category IDs can
            be looked up via /open-api/{lang}/Miscellaneous/Categories?type=Tours.
          schema:
            type: string
          example: "532"
        - name: page
          in: query
          required: false
          description: Page number; upstream returns up to 30 results per page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        "200":
          description: Themed tours retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeResponse'
        "400":
          description: Unsupported language
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToursErrorResponse'
              example:
                error: "unsupported language: xx"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToursErrorResponse'
              example:
                error: "upstream service unavailable"

  # Taipei Travel miscellaneous proxy
  /open-api/{lang}/Miscellaneous/Categories:
    get:
      tags:
        - Miscellaneous
      summary: List categories for a given resource type
      description: |
        Compatibility proxy for the Taipei Travel Miscellaneous/Categories
        open API.

        No authentication is required. The "type" query parameter is
        required by the upstream API; this backend validates it before
        forwarding the request.
      security: []
      parameters:
        - name: lang
          in: path
          required: true
          description: Response language
          schema:
            type: string
            enum:
              - zh-tw
              - zh-cn
              - en
              - ja
              - ko
          example: zh-tw
        - name: type
          in: query
          required: true
          description: Resource type to look up categories for
          schema:
            type: string
            enum:
              - Activity
              - Calendar
              - Pictorial
              - Attractions
              - Accommodation
              - Tours
          example: Tours
      responses:
        "200":
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'
        "400":
          description: Unsupported language, or missing/invalid "type" parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiscellaneousErrorResponse'
              examples:
                missing_type:
                  summary: Missing "type" parameter
                  value:
                    error: "invalid parameter: type ()"
                invalid_type:
                  summary: Invalid "type" value
                  value:
                    error: "invalid parameter: type (Foo)"
        "502":
          description: Taipei Travel upstream API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiscellaneousErrorResponse'
              example:
                error: "upstream service unavailable"